fix: integrate runtime with application
The application was not correctly starting the runtime in the
application tree.

But, it was attempting to start it before the LSP had initialized, so we
instead switch to starting the runtime dynamically under a
DynamicSupervisor.

This also required having to set the group leader on the remote node to
the `:user` process and manually compiling the runtime's deps and
project.